refactor: DTO 구조 개선 - 도메인별 dto.py 분리
- types/ 모듈 제거 (순환 참조 방지용 임시 모듈)
- retrieval/dto.py 신설: QueryPlan, SearchResult, ExpandedResult, SelfQueryResult
- generation/dto.py 신설: models.py를 dto.py로 rename
- ingestion/dto.py 신설: models.py를 dto.py로 rename
- 각 도메인 간 데이터 전송 객체를 해당 도메인 내 dto.py에서 관리
- Spring/Java DTO 패턴을 Python 프로젝트에 적용
