+++ poc/src/rag/cli/use_cases/search.py
@@ -6,7 +6,7 @@ Orchestrates the retrieval pipeline for search operations.
 from typing import List, Optional, Protocol
 
 from src.rag.retrieval import RetrievalPipeline
-from src.rag.types import ExpandedResult
+from src.rag.retrieval.dto import ExpandedResult
 from src.rag.shared.config import EmbeddingConfig
